[debops.ferm] Enable ferm on hosts without cap12s
In case that Ansible didn't detect any POSIX capabilities, assume by
default that the host is a normal hardware server/VM and firewall
configuration is possible. This should avoid idempotency issues when on
the first run 'libcap2-bin' package is not installed and on a second run
it suddenly shows up, which prompts the 'debops.ferm' role to configure
ferm, breaking idempotency.